requirements.txt
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.0
meteostat==1.6.8
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
psutil==5.9.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.10
pytz==2025.1
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1
